Fixed MAX log files type.

+++ data.go
@@ -43,7 +43,7 @@ var logfileNameList []string
 var dummyLogfile string
 //var loggerWG sync.WaitGroup
 
-const log_MAX_FILES int = 10
+const log_MAX_FILES int8 = 10
 const log_FILE_NAME_PREFIX string = "server.log"
 const default_LOG_LEVEL string = "debug"
 const log_FILE_SIZE int64 = 20971520 // 20 MB
